changed float to float64 in regret matching

+++ poker_solver/deep_poker_solver.py
@@ -254,11 +254,11 @@ class DeepPokerSolver:
         positive_regrets_sum = torch.sum(regrets[positive_regrets])
         if positive_regrets_sum < 1e-8:
             return (
-                torch.ones(regrets.shape, device=self.device, dtype=torch.float)
+                torch.ones(regrets.shape, device=self.device, dtype=torch.float64)
                 / regrets.shape[0]
             )
         
-        strategy = torch.zeros(regrets.shape, device=self.device, dtype=torch.float)
+        strategy = torch.zeros(regrets.shape, device=self.device, dtype=torch.float64)
         strategy[positive_regrets] += regrets[positive_regrets] / positive_regrets_sum
         return strategy
